feat: render branded auth on tenant subdomain (issue #168)
Extract SignInForm into components/auth/sign-in-form.tsx with optional
defaultSlug/tenantName/logoUrl props. Replace the tenant sign-in redirect
with a server component that fetches tenant branding and renders the form
in place. Magic-link confirm URLs always target the root domain since the
/auth/confirm handler only exists on the platform routes.

Update E2E auth tests to reflect that the sign-in page now renders on the
tenant subdomain (no redirect) and that the password field requires clicking
"Use Password" to reveal before filling.
